# mes-root/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Mes Auto
  description: auto api
  contact:
    email: ixtf1984@gmail.com
  version: 1.0.0
security:
- bearerAuth: []
...
paths:
...
  /api/admin/silkCarRecords/{id}:
    delete:
      operationId: handle_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json: {}
...
  /api/autoComplete/silkCar:
    get:
      operationId: silkCar
      parameters:
      - name: q
        in: query
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: object
...
  /api/packageBoxes/{id}/silks:
    get:
      operationId: silks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Silk'
  /api/packageBoxes/{id}/silkCarRecords:
    get:
      operationId: silkCarRecords
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SilkCarRecord'
...
  /api/silkCarRecords/{id}/events:
    get:
      operationId: events
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventSource'
  /api/silkCarRecords:
    get:
      operationId: query_10
      parameters:
      - name: first
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
          default: 50
      - name: silkCarCode
        in: query
        schema:
          type: string
      - name: endDate
        in: query
        schema:
          type: string
      - name: startDate
        in: query
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: object
...
  /api/silkCars/{id}:
    get:
      operationId: get_23
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SilkCar'
    put:
      operationId: update_15
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SilkCar'
  /api/silkCars:
    get:
      operationId: query_11
      parameters:
      - name: first
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
          default: 50
      - name: q
        in: query
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: object
    post:
      operationId: create_18
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SilkCar'
...
  /api/silks/{id}/silkCarRecords:
    get:
      operationId: silkCarRecords_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SilkCarRecord'
  /api/silks/{id}/packageBox:
    get:
      operationId: packageBox
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageBox'
...
  /api/silks/{id}:
    get:
      operationId: get_27
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Silk'
  /api/silks:
    get:
      operationId: query_12
      parameters:
      - name: first
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
          default: 50
      - name: batchId
        in: query
        schema:
          type: string
      - name: workshopId
        in: query
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
...
    Batch:
      required:
      - batchNo
      - id
      - product
      - spec
      - tubeColor
      - workshop
      type: object
      properties:
        id:
          type: string
        workshop:
          $ref: '#/components/schemas/Workshop'
        product:
          $ref: '#/components/schemas/Product'
        batchNo:
          type: string
        silkWeight:
          minimum: 1
          type: number
          format: double
        centralValue:
          minimum: 1
          type: number
          format: double
        holeNum:
          minimum: 1
          type: integer
          format: int32
        spec:
          type: string
        tubeColor:
          type: string
        note:
          type: string
    Corporation:
      required:
      - code
      - id
      - name
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
    DyeingResult:
      required:
      - createDateTime
      - creator
      - dateTime
      - id
      - lineMachine
      - silk
      type: object
      properties:
        id:
          type: string
        silk:
          $ref: '#/components/schemas/Silk'
        lineMachine:
          $ref: '#/components/schemas/LineMachine'
        spindle:
          minimum: 1
          type: integer
          format: int32
        dateTime:
          type: string
          format: date-time
        hasException:
          type: boolean
        grade:
          $ref: '#/components/schemas/Grade'
        silkExceptions:
          type: array
          items:
            $ref: '#/components/schemas/SilkException'
        silkNotes:
          type: array
          items:
            $ref: '#/components/schemas/SilkNote'
        creator:
          $ref: '#/components/schemas/Operator'
        createDateTime:
          type: string
          format: date-time
        modifier:
          $ref: '#/components/schemas/Operator'
        modifyDateTime:
          type: string
          format: date-time
        submitted:
          type: boolean
        formConfigValueData:
          $ref: '#/components/schemas/JsonNode'
        formConfig:
          $ref: '#/components/schemas/JsonNode'
    DyeingResultInfo:
      type: object
      properties:
        dyeingResult:
          $ref: '#/components/schemas/DyeingResult'
        grade:
          $ref: '#/components/schemas/Grade'
    EventSource:
      type: object
      properties:
        eventId:
          type: string
        operator:
          $ref: '#/components/schemas/Operator'
        fireDateTime:
          type: string
          format: date-time
        deleted:
          type: boolean
        deleteOperator:
          $ref: '#/components/schemas/Operator'
        deleteDateTime:
          type: string
          format: date-time
        type:
          type: string
          enum:
          - SilkCarRuntimeInitEvent
          - SilkCarRuntimeAppendEvent
          - SilkCarRuntimeGradeEvent
          - SilkRuntimeDetachEvent
          - SilkCarRuntimeGradeSubmitEvent
          - DyeingSampleSubmitEvent
          - TemporaryBoxEvent
          - ProductProcessSubmitEvent
          - ExceptionCleanEvent
          - DyeingPrepareEvent
          - PackageBoxEvent
          - PackageBoxFlipEvent
          - SilkNoteFeedbackEvent
          - JikonAdapterSilkCarInfoFetchEvent
          - JikonAdapterSilkDetachEvent
          - JikonAdapterPackageBoxEvent
          - RiambPackageBoxEvent
          - RiambSilkCarInfoFetchEvent
          - RiambSilkDetachEvent
          - WarehousePackageBoxFetchEvent
    Grade:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        sortBy:
          type: integer
          format: int32
    JsonNode:
      type: object
      properties:
        valueNode:
          type: boolean
        containerNode:
          type: boolean
        missingNode:
          type: boolean
        object:
          type: boolean
        pojo:
          type: boolean
        integralNumber:
          type: boolean
        floatingPointNumber:
          type: boolean
        textual:
          type: boolean
        boolean:
          type: boolean
        binary:
          type: boolean
        int:
          type: boolean
        number:
          type: boolean
        bigDecimal:
          type: boolean
        bigInteger:
          type: boolean
        double:
          type: boolean
        long:
          type: boolean
        float:
          type: boolean
        nodeType:
          type: string
          enum:
          - ARRAY
          - BINARY
          - BOOLEAN
          - MISSING
          - NULL
          - NUMBER
          - OBJECT
          - POJO
          - STRING
        short:
          type: boolean
        array:
          type: boolean
        null:
          type: boolean
    Line:
      required:
      - id
      - name
      - workshop
      type: object
      properties:
        id:
          type: string
        workshop:
          $ref: '#/components/schemas/Workshop'
        name:
          type: string
        doffingType:
          type: string
          enum:
          - AUTO
          - MANUAL
          - DYEING_SAMPLE
          - PHYSICAL_INFO
    LineMachine:
      required:
      - id
      - line
      - spindleSeq
      type: object
      properties:
        id:
          type: string
        line:
          $ref: '#/components/schemas/Line'
        item:
          minimum: 1
          type: integer
          format: int32
        spindleNum:
          minimum: 1
          type: integer
          format: int32
        spindleSeq:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            type: integer
            format: int32
    Operator:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        hrId:
          type: string
        oaId:
          type: string
        phone:
          type: string
        admin:
          type: boolean
    Product:
      required:
      - code
      - id
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
    SapT001l:
      required:
      - id
      type: object
      properties:
        id:
          type: string
        lgort:
          type: string
        lgobe:
          type: string
    Silk:
      required:
      - batch
      - code
      - doffingDateTime
      - doffingOperator
      - doffingType
      - grade
      - id
      - lineMachine
      - spindle
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        doffingNum:
          type: string
        doffingOperator:
          $ref: '#/components/schemas/Operator'
        doffingType:
          type: string
          enum:
          - AUTO
          - MANUAL
          - DYEING_SAMPLE
          - PHYSICAL_INFO
        doffingDateTime:
          type: string
          format: date-time
        lineMachine:
          $ref: '#/components/schemas/LineMachine'
        spindle:
          type: integer
          format: int32
        batch:
          $ref: '#/components/schemas/Batch'
        grade:
          $ref: '#/components/schemas/Grade'
        exception:
          $ref: '#/components/schemas/SilkException'
        exceptions:
          type: array
          items:
            $ref: '#/components/schemas/SilkException'
        dyeingExceptionStrings:
          type: array
          items:
            type: string
        dyeingSample:
          type: boolean
        detached:
          type: boolean
        packageDateTime:
          type: string
          format: date-time
    SilkCar:
      required:
      - code
      - id
      - number
      - type
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum:
          - DEFAULT
          - BIG_SILK_CAR
        number:
          type: string
        code:
          type: string
        row:
          minimum: 1
          type: integer
          format: int32
        col:
          minimum: 1
          type: integer
          format: int32
    SilkCarRecord:
      required:
      - batch
      - grade
      - id
      type: object
      properties:
        id:
          type: string
        silkCar:
          $ref: '#/components/schemas/SilkCar'
        batch:
          $ref: '#/components/schemas/Batch'
        grade:
          $ref: '#/components/schemas/Grade'
        doffingOperator:
          $ref: '#/components/schemas/Operator'
        doffingType:
          type: string
          enum:
          - AUTO
          - MANUAL
          - DYEING_SAMPLE
          - PHYSICAL_INFO
        doffingDateTime:
          type: string
          format: date-time
        carpoolOperator:
          $ref: '#/components/schemas/Operator'
        carpoolDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
        startDateTime:
          type: string
          format: date-time
        initCommand:
          $ref: '#/components/schemas/JsonNode'
        initSilks:
          type: array
          items:
            $ref: '#/components/schemas/SilkRuntime'
...
    SilkException:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    SilkNote:
      required:
      - id
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mustFeedback:
          type: boolean
    SilkRuntime:
      type: object
      properties:
        sideType:
          type: string
          enum:
          - A
          - B
        row:
          type: integer
          format: int32
        col:
          type: integer
          format: int32
        silk:
          $ref: '#/components/schemas/Silk'
        exceptions:
          type: array
          items:
            $ref: '#/components/schemas/SilkException'
        notes:
          type: array
          items:
            $ref: '#/components/schemas/SilkNote'
        grade:
          $ref: '#/components/schemas/Grade'
        firstDyeingResultInfo:
          $ref: '#/components/schemas/DyeingResultInfo'
        crossDyeingResultInfo:
          $ref: '#/components/schemas/DyeingResultInfo'
        multiDyeingResultInfo:
          $ref: '#/components/schemas/DyeingResultInfo'
    Workshop:
      required:
      - code
      - corporation
      - id
      - name
      type: object
      properties:
        id:
          type: string
        corporation:
          $ref: '#/components/schemas/Corporation'
        name:
          type: string
        code:
          maxLength: 1
          minLength: 1
          type: string
        note:
          type: string
        sapT001ls:
          type: array
          items:
            $ref: '#/components/schemas/SapT001l'
        sapT001lsForeign:
          type: array
          items:
            $ref: '#/components/schemas/SapT001l'
        sapT001lsPallet:
          type: array
          items:
            $ref: '#/components/schemas/SapT001l'
...
    PackageBox:
      required:
      - batch
      - code
      - grade
      - id
      - type
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum:
          - AUTO
          - MANUAL
          - MANUAL_APPEND
          - WORK_FETCH
          - FOREIGN
        code:
          type: string
        saleType:
          type: string
          enum:
          - DOMESTIC
          - FOREIGN
        palletCode:
          type: string
        palletType:
          type: string
        packageType:
          type: string
        foamType:
          type: string
        foamNum:
          type: integer
          format: int32
        batch:
          $ref: '#/components/schemas/Batch'
        grade:
          $ref: '#/components/schemas/Grade'
        silkCount:
          minimum: 1
          type: integer
          format: int32
        grossWeight:
          minimum: 1
          type: number
          format: double
        netWeight:
          minimum: 1
          type: number
          format: double
        riambJobId:
          type: string
        printDate:
          type: string
          format: date-time
        printClass:
          $ref: '#/components/schemas/PackageClass'
        printCount:
          minimum: 0
          type: integer
          format: int32
        automaticPackeLine:
          type: string
        budat:
          type: string
          format: date-time
        budatClass:
          $ref: '#/components/schemas/PackageClass'
        sapT001l:
          $ref: '#/components/schemas/SapT001l'
        inWarehouse:
          type: boolean
        creator:
          $ref: '#/components/schemas/Operator'
    PackageClass:
      required:
      - id
      - name
      - riambCode
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        riambCode:
          type: string
        sortBy:
          type: integer
          format: int32
...